fix: keep campaign validators smell free
